+++ crates/block-producer/src/store/mod.rs
@@ -6,20 +6,20 @@ use std::{
 
 use itertools::Itertools;
 use miden_node_proto::{
-    domain::note::NoteAuthenticationInfo,
+    domain::batch::BatchInputs,
     errors::{ConversionError, MissingFieldHelper},
     generated::{
         digest,
         requests::{
-            ApplyBlockRequest, GetBlockHeaderByNumberRequest, GetBlockInputsRequest,
-            GetNoteAuthenticationInfoRequest, GetTransactionInputsRequest,
+            ApplyBlockRequest, GetBatchInputsRequest, GetBlockHeaderByNumberRequest,
+            GetBlockInputsRequest, GetTransactionInputsRequest,
         },
         responses::{GetTransactionInputsResponse, NullifierTransactionInputRecord},
         store::api_client as store_client,
     },
     AccountState,
 };
-use miden_node_utils::formatting::format_opt;
+use miden_node_utils::{formatting::format_opt, tracing::grpc::OtelInterceptor};
 use miden_objects::{
     account::AccountId,
     block::{Block, BlockHeader, BlockNumber},
@@ -29,7 +29,7 @@ use miden_objects::{
     Digest,
 };
 use miden_processor::crypto::RpoDigest;
-use tonic::transport::Channel;
+use tonic::{service::interceptor::InterceptedService, transport::Channel};
 use tracing::{debug, info, instrument};
 
 use crate::{block::BlockInputs, errors::StoreError, COMPONENT};
@@ -121,17 +121,19 @@ impl TryFrom<GetTransactionInputsResponse> for TransactionInputs {
 // STORE CLIENT
 // ================================================================================================
 
+type InnerClient = store_client::ApiClient<InterceptedService<Channel, OtelInterceptor>>;
+
 /// Interface to the store's gRPC API.
 ///
 /// Essentially just a thin wrapper around the generated gRPC client which improves type safety.
 #[derive(Clone)]
 pub struct StoreClient {
-    inner: store_client::ApiClient<Channel>,
+    inner: InnerClient,
 }
 
 impl StoreClient {
     /// TODO: this should probably take store connection string and create a connection internally
-    pub fn new(store: store_client::ApiClient<Channel>) -> Self {
+    pub fn new(store: InnerClient) -> Self {
         Self { inner: store }
     }
 
@@ -212,21 +214,17 @@ impl StoreClient {
     #[instrument(target = COMPONENT, skip_all, err)]
     pub async fn get_batch_inputs(
         &self,
+        block_references: impl Iterator<Item = (BlockNumber, Digest)> + Send,
         notes: impl Iterator<Item = NoteId> + Send,
-    ) -> Result<NoteAuthenticationInfo, StoreError> {
-        let request = tonic::Request::new(GetNoteAuthenticationInfoRequest {
+    ) -> Result<BatchInputs, StoreError> {
+        let request = tonic::Request::new(GetBatchInputsRequest {
+            reference_blocks: block_references.map(|(block_num, _)| block_num.as_u32()).collect(),
             note_ids: notes.map(digest::Digest::from).collect(),
         });
 
-        let store_response =
-            self.inner.clone().get_note_authentication_info(request).await?.into_inner();
+        let store_response = self.inner.clone().get_batch_inputs(request).await?.into_inner();
 
-        let note_authentication_info = store_response
-            .proofs
-            .ok_or(GetTransactionInputsResponse::missing_field("proofs"))?
-            .try_into()?;
-
-        Ok(note_authentication_info)
+        store_response.try_into().map_err(Into::into)
     }
 
     #[instrument(target = COMPONENT, skip_all, err)]