+++ crates/block-producer/src/test_utils/mod.rs
@@ -1,5 +1,7 @@
 use miden_objects::{
+    account::AccountId,
     crypto::rand::{FeltRng, RpoRandomCoin},
+    testing::account_id::AccountIdBuilder,
     transaction::TransactionId,
     Digest,
 };
@@ -43,6 +45,10 @@ impl Random {
 
     pub fn draw_tx_id(&mut self) -> TransactionId {
         self.0.draw_word().into()
+    }
+
+    pub fn draw_account_id(&mut self) -> AccountId {
+        AccountIdBuilder::new().build_with_rng(&mut self.0)
     }
 
     pub fn draw_digest(&mut self) -> Digest {